requirements.txt
# Web app
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
websockets>=12.0

# Google Cloud
google-auth>=2.0.0
google-cloud-secret-manager>=2.16.0
google-cloud-aiplatform>=1.38.0
google-api-python-client>=2.100.0
